refactor(DatasourceEditor): extract test utilities and add factory function
Addresses PR feedback to improve test isolation and maintainability:

1. **Extract shared utilities** - Created DatasourceEditor.test.utils.tsx to
   isolate shared test fixtures and functions (props, DATASOURCE_ENDPOINT,
   asyncRender, setupDatasourceEditorMocks). This fixes import coupling that
   caused Currency tests to execute all 13 DatasourceEditor tests (16 total
   instead of 3).

2. **Add factory function** - Converted propsWithCurrency to
   createPropsWithCurrency() factory that returns fresh deep copies on each
   call. This prevents potential test pollution from shared mutable fixtures.

Results:
- Currency tests now run 3 tests in 79s (was 16 tests in 175s - 55% faster)
- All 40 DatasourceEditor tests passing
- Better test isolation and maintainability
